Host-agent implementation for CloudFoundry
This change implements CfEnvironment to allow
the agent to run in ClodFoundry deplopyments.
The etcd key-value store available in CloudFoundry
is used as the basis for creating EP and Service
files; this information is written by the
controller and the agent mostly reacts by creating
the appropriate EP and Service files.
To support legacy CloudFoundry networking
(non-CNI networking), the agent also configures
iptables rules.

Signed-off-by: Amit Bose <amitbose@gmail.com>