[CloudFoundry] Controller high availability
Add support for simple high availability of the
controller by allowing multiple instances of it
to run in single-master mode. Master/slave role
is managed through an external lock provided by
the Locket component in CF. At start-up, the
controller that grabs a lock successfully becomes
the master and actively updates ACI and etcd.
Otherwise the controller just pauses during start-up
and polls Locket for the lock.

Signed-off-by: Amit Bose <amitbose@gmail.com>