[CloudFoundry] Introduce key-value service
Package keyvalueservice is a library that provides
a remotely accessible in-memory versioned key-value store.
This will be used to communicate data/events between the
controller and host-agent (much like the way etcd is
being used). In fact, the goal of this library is to
provide a replacement for etcd. This key-value store
is not a generic replacement for etcd - it provides
only those features that are currently being used by
the controller/host-agent.

The key-value store is accessed locally using struct
KvManager - it organizes keys by "namespaces". There
is no hierarchical relationship between keys within
a namespaces. In addition to set, get, delete, one
can also watch for changes to the store. The manager
is made available over gRPC using the KvService struct.
KvWatcher is a helper class that uses the RPC service
to fetch changes from a remote key-value store.
KvServer and KvClient are wrappers for the server and
client side that deal with handling connections.