Avoid using stale opflex device MAC
When nodes are re-created, APIC creates
a brand new opflex-device object for it. This
new object and the old object are both present
in the nodeOpflexDevice map till the old object
ages out on APIC and gets deleted from the map.
In this fix, deviceMacForNode() is changed to use
the newer opflex device instead of the stale one.

Signed-off-by: Amit Bose <amitbose@gmail.com>