Fix net namespace configuration for non-root Pods
When K8S pods are configured to run as non-root
users, the host-agent is denied access to that
pods network namespace (even though the agent is
runnng as root). This fix switches the file-system
user-id of the child-mode host-agent to the pod's
user-id as a work-around.

Signed-off-by: Amit Bose <amitbose@gmail.com>
(cherry picked from commit f1a564c9abdf320eaf3af3654bbd5c02c092ce96)