Change the APIC refresh Timer to be configurable.
As a part of this change we have a added a new parameter in the
controller configMap.
This parameter will determine the refresh rate at which the the
client will refresh with the APIC.
If no value given then the default is taken.

(cherry picked from commit 60018bcbab19f9c7550d7fe7b11b5515ab11dff5)