Default APIC websocket subscription refresh timeout
The apic-refreshtime configuration was introduced in the commit:
60018bcbab19f9c7550d7fe7b11b5515ab11dff5

The current commit updates the default value of that configuration
to a larger value.

Also clarifying comments added as to what the different values mean
for this configuration, namely:

* not providing this configuration in the configuration file will
default it's value to 900 seconds
* setting it to zero will result in APIC's default being used (which
currently is 80 secs)
* setting to a value greater than 0 will result in that value
being used in the APIC connection.

(cherry picked from commit fcb6e868f812b6234535454b0a10d5175c362c0d)