Option to add contract scope for k8s services as an annotation
Default value: context

For example, add in the service definition:
"metadata": {
  "annotations": {
    "opflex.cisco.com/ext_service_scope" : "tenant"
  }
}

Valid scope values: context, tenant, global. An empty entry defaults to context(VRF only)

If a new service is deployed with invalid scope, it wont register with APIC
If an existing service is updated with invalid scope, the service wont be updated in APIC

Added unit and functional tests to check the contract scope values and service annotations.
Also formatted the committed files with go fmt