kubernetes 12 nodes come with default taints
Taints:             node-role.kubernetes.io/master:NoSchedule
                    node.kubernetes.io/not-ready:NoSchedule

As a result aci daemonsets cannot start with the current tolerations
- key: CriticalAddonsOnly
- effect: NoSchedule
  key: node-role.kubernetes.io/master

We change this to tolerate any key with effect NoSchedule.

This is the same approach used by weave and other cni plugins.

Signed-off-by: Madhu Challa "challa@noironetworks.com"

Put a "-" in front of effect to keep prov happy.

Signed-off-by: Madhu Challa "challa@noironetworks.com"

Revert previous change "-" and fix test files.

Signed-off-by: Madhu Challa "challa@noironetworks.com"