Option to enable security-sharing in ext subnets for k8s services
Default value: import-security

If a service contract is defined with the scope value "global", the respective extSubnet is registered with shared-security scope setting.
The extSubnet scope thus becomes: "import-security, shared-security"

This is an addition to https://github.com/noironetworks/aci-containers/pull/186 where we first added option to change service contract scope

Also modified the unit test to verify contract scope to include this code change. TestServiceAnnotation now also checks for the extSubnet's scope value when the service contract scope is set to global