IPSLA handling for SNAT
When IPSLA is enabled (by configuring IPSLA interval to a >0 value), create one health-group per compute-node.
This will make sure that from fabric side, the packets are not hashed/black-holed towards the compute-nodes that are not active.

Fix has been tested on vk8s-2 where 3 of the nodes doesn't have any user-pods (ARP flows are not configured).
Also, tried shrinking k8s cluster and observed that the health-group is cleaned-up accordingly.