Fix tagAnnotation dn format
It appears that APIC4.1 onwards seems to have changed the format of tagAnnotation object.
This change will accommodate the same, so that it avoids treating an object as altered when processing a websocket notification on any of the ACC managed objects.

Example:
Prior to 4.1: uni/tn-common/brc-vk8s_2_snat_svcgraph/annotationKey-aci-containers-controller-tag
4.1+        : uni/tn-common/brc-vk8s_2_snat_svcgraph/annotationKey-[aci-containers-controller-tag]

Tests:
Reproduced the issue on Fab14 and tested with fix.