Configure snat namespace for hostagent nodeinfo processing
aci-snat-namespace can be set as field in host-agent-config in the aci deployment file. Set to "aci-containers-system" by default.
If the user sets this field to a garbage value(i.e not a valid k8s namespace), print an error statement and ignore. The user needs to either remove that namespace value or set it to something valid.

(cherry picked from commit 0ec5d07233f370cd11158bf08ce33caab446ac15)