Change null map ep file exists check
Currently host-agent pod creates a null mac ep file only if there's no file with "null-mac" in its name.
There might a situation in dev/prod mode where we provision an existing cluster with new naming convention. We need to be able to delete the "kube-default_null-mac" ep file and create "aci-containers-default_null-mac" file instead.
To accomplish that, we now look for the entire filename with default EPG name in it.