Host-agent configures directly Uplink-mac address as Service-mac
Opflex-agent  inject the Uplink MAC address to APIC, which in turn  ACC controllers reads that Mac from APIC by subscribing to OpflexDevice change,
and annotates that using node info,  and host-agent uses that Mac to program the Service-mac in file.
If opflex-agent fails to inject that device  Mac,  acc controller allocates a random Mac. Because of this random Mac, inconsistency is coming between interface  Mac and service mac.

with SnatPolicy installed OVS adds a rule to allow only packets destained to the interface Mac this is breaking the Service and Snat traffic.
Now to fix this issue host-agent directly uses the Uplink MAC address to configure Service-mac in service files