Fixed  Session Affinity stickiness issue when session affinity config is  removed
This issue is observed when the session affinity config is added and removed with large timeouts.
After removing the session affinity kernel flow is not getting timed-out, due to opflex-agent do not cache the old flows to get the diff out of it
solution is to set the timeout to be 1 sec. so that flow will be expired

(cherry picked from commit 50bac87fafaa7bdf4bcdf911f17e0d8dbb3fafff)