Changes for acc-provision-operator
A new operator is introduced for cluster upgrades and lifecycle management. This requires the acicnioperator pod to read acc-provision configmap and create a AccProvisionInput CR based on the unmarshalled data.
Relevant files changed/added:
1. cmd/acicontainersoperator/main.go: Add a REST client for acc-provision operator
2. pkg/accprovisioninput/: New package with the acc-provision CRD definition used by the acicnioperator to construct the CR
3. pkg/controller/acicontainersoperator.go: Add code to read the acc-provision configmap and create CR

Additionally fix a small bug in the Status field definition for acicnicontainersoperator CRD and fix gofmt on pkg/hostagent