Enabling wait-for-network for pods
This check allows the CNI binary to check if the
pod's datapath is working before it successfully
returns from the call made by the kubelet to add
an interface for the pod.

The check was already present, this change enables
it by default. It can be disabled by setting the
net_config/disable_wait_for_network configuration
to True.