Fix bug in SNAT policy computation logic
Currently we compute IP and port allocations for every policy when there's a change in any snat policy. This causes us to allocate additional port ranges to a node and policy pair. To avoid this, we want to skip computing if the policy is already applied on that node.

Also added more logging on INFO and DEBUG levels for SNAT

(cherry picked from commit 8b50ea2e9893704a9481ad09cfb71d89d241de64)