Update policy status from IpPortExhausted to Ready after a node is scaled down
When the policy status is IPPortExhausted and cannot update to ready state when a node is deleted,
we iterate through the snatPolicyNames and check the status for each policy.
This is because when there are no more ports that can be allocated, there is no entry of that node
in the global info cache.

Signed-off-by: Tanya Tukade tanyatukade.123@gmail.com