Handling delete service by adding in queue
Sometimes preparing the apic object and updating cache was happening for
delete ahead of update. This was because handling of delete was done
immediately and that of update was done by adding in queue.

Modified code to process the delete also in queue.

(cherry picked from commit 3b87c4781bf0b2e846d3c0ba8f649df4ea6b1a93)