Cluster Report : Collect all available logs files of a pod
At the moment cluster-report collect only the last (running) log
for our CNI components. Very often this is not enough if we have
to do RCA for an issue that happened several days ago and we
have to manually collect the rest. We need the the cluster-report
to collect *ALL* available log files in:

/var/log/pods/aci-containers-system_aci-containers-host-*/opflex-agent/*
/var/log/pods/aci-containers-system_aci-containers-host-*/mcast-daemon/*
/var/log/pods/aci-containers-system_aci-containers-host-*/aci-containers-host/*
/var/log/pods/aci-containers-system_aci-containers-openvswitch-*/aci-containers-openvswitch/*
/var/log/pods/aci-containers-system_aci-containers-controller-*/aci-containers-controller/*
/var/log/pods/aci-containers-system_aci-containers-operator-*/aci-containers-operator/*