Fix controller not switching to next APIC
Root cause:
This occurs when the controller comes up and the first APIC is
already down.

The controller tries to get the APIC version, and in the GetVersion
method when the APIC login fails we keep on retrying the login
hence the rollover is not happening.

Fix:
Add retry limit and rollover to the next APIC.