Periodically check and remove the taint added by the controller
Issue: The taint added by controller when the node is not ready
is removed by the host agent only once during it's initialization,
but if the node becomes not ready after the host agent is running,
the controller will add the taint again and as the host agent will
already be running the taint will not be removed.

Fix: Instead of removing the taint and exiting, made goroutine to
check periodically for taint, if present and node in ready state,
remove the taint.