Sort nodes selected for snat service redirect policy
If number of nodes in the cluster is greater than the maximum number of
nodes supported for service graph, the nodes selected for the service
graph might be different each time if we don't sort the nodes.

Sorted the nodes based on name of node to avoid having a different node
list each time.