Fix Service with dynamic allocation takes IP out of static pool
Root Cause:
The IP of statically allocated service was getting added to the
list of ingress IPs which are dynamically allocated, and from
there added to the list of avaialble service IPs, so this new
added IP was allocated to the service with dynamic allocation.

Added check to add only the dynamic allocation IPs to the
corresponding list.