Fix APIC sync not complete when hpp rendering is disabled
Issue and Root cause:
When disable-hpp-rendering config flag is set to true, processing
of the netPolQueue is skipped. Before APIC sync, we put a channel
into all the work queues and wait on these channels to check if
all the work queues are processed. As netPolQueue is not processed
when disable-hpp-rendering is true, the channel is not closed
causing the process to wait forever and the checkpoint will not
be completed, so APIC sync will not be complete.

Fix:
Skipped adding channel to netPolQueue when disable-hpp-rendering
config flag is set to true.