Fix failed to update service status error
Error:
msg="Failed to update service status: Operation cannot be fulfilled
on services <name>: the object has been modified; please apply your
changes to the latest version and try again"

Root cause:
The service was updated twice once with the IP and again with the
condition without getting the updated service.

Fix:
As the service was updated back to back, made changes to do the
update only once with both the IP and condition status.