Fix Unit test TestSnatnodeInfo failing randomly
After snat IP update the ports are not expected to be same as
before the update. In the invalid nodeInfo delete test which is
after the snat IP update test, the port ranges expected to be same
as before the update, hence the test case was failing randomly
sometimes.

Moved the invalid delete tests before the update test.