Fix opflexSnatLocalInfos cache deletion logic for pod
Issue: When multiple SNAT policies are continuously created and
deleted, the snatlocalinfos CRs sometimes fail to be created for
few nodes.

Root Cause: The opflexSnatLocalInfos for a pod was being deleted
within the updateEpFiles function if no SNAT policy was selected
for the pod. This function is triggered by both snatpolicy and
snatglobalinfos events. In some cases, when a SNAT policy is
deleted and recreated, the sequence of create and delete events
for snatpolicy and snatglobalinfos can become unordered.
This results in the deletion event of the previous SNAT policy
occurring after the updation event of the snatglobalinfos for
current policy, causing the current cache entry to be deleted
after it is created. Consequently, the opflexSnatLocalInfos is
not created due to a mismatch between opflexSnatLocalInfos and
opflexSnatGlobalInfos.

Fix: The deletion of opflexSnatLocalInfos for a pod has been moved
to function deleteSnatLocalInfo, which is triggered only by
snatpolicy events and not by snatglobalinfos events.