Handled network policies with CIDR for hpp direct
When enable_hpp_direct was enabled, network policies with CIDR were
not working as expected as the CIDR was not added to the hostprotpol CR
and thus not added to the netpol file provided by host-agent container
to the opflex-agent.
Added code to handle it.

(cherry picked from commit b681da881d1acbe3dec66158888e871e244fd2cb)