Fix to delete stale NodeInfo CR
When a node is removed from a cluster for which NodeInfo CR was present
with snatpolicies, controller was deleting the NodeInfo CR for the node,
but hostagent pod for the deleted node was remaining in the cluster for
sometime even after the node was removed and it was re-creating the
NodeInfo CR and SnatLocalInfo CR which remained as stale resources
because of which snatpolicies were in IpPortsExhausted state.

Added fix not to re-create NodeInfo or SnatLocalInfo Cr when node is not
present in cluster and delete the stale CRs that are already present in
the cluster.

(cherry picked from commit 7598f897a34053e89b7e8e4532d6b554996f89c8)