Delete stale snatlocalinfo CRs
When a node with applied snatpolicies is removed from the OCP cluster,
only the nodeinfo CR was deleted, while the snatlocalinfo CR remained.

This commit modifies the code to ensure that snatlocalinfo CRs are also
deleted when the corresponding node is removed. Additionally, the code
is updated to remove all snatlocalinfo CRs in the cluster for nodes
and nodeinfos that have already been deleted.

(cherry picked from commit 0e9d03faa31185b03774d6423b9174ecf8fb790e)