Fix handling of HostNetwork pods in hpp_direct
When hpp_direct is enabled, the hostprotremoteipcontainer
CR was previously storing a single entry for each IP address,
which caused issues when multiple pods with the same IP but
different labels existed in the same namespace. Deleting
one pod would result in the IP being removed from the
hostprotremoteipcontainer CR, even if other pods using the
same IP remained.

This update modifies the code to add a separate entry for
each pod in the hostprotremoteipcontainer CR, ensuring
that IPs are only removed when all pods using the IP have
been deleted.