Limit opflexODev processing to cluster vmmDomain
Since the controller subscribes to the entire opflexODev class, it
receives notifications for opflexODev objects from all clusters.
Each modified update triggers a GET operation to APIC, causing
increased load.

To avoid this, the opflexODev objects corresponding to the cluster
vmmDomain are stored in the cache. Updates for opflexODev objects
not present in the cache are skipped, preventing unnecessary GET
operations to APIC for objects that don't belong to the cluster.