Ensure correct processing of namespace labels for snat
An snat policy with a label selector would not be applied to a Namespace object because the matching logic was incorrectly using the empty `.metadata.namespace` field instead of the object's name.

This change updates the logic to use the `.metadata.name` field when evaluating a Namespace, allowing policies to match as expected.

(cherry picked from commit 0cb378342726a45f2332e3bd765539f904e03606)