Read PlatformConfigDelete event from opflex.
With this change, host-agent subscribes for the PlatformConfig delete
notifications from opflex-agent.
Opflex-agent will update the notification file on PlatformConfig delete event.
This notification will be read by host-agent when there is a node-changed
event triggered by controller. If this notification contains timestamp which is
less than 2mins old then it considers as an intra-pod migration and
skips the dhcp release-renew steps.