Remove stale opflexOdevs
Sometimes after TOR reboots, its observed that we don't receive
opflexOdev delete and thus its entry remains in controller's cache.
This stale cache affects the way we get aci-annotation for the node.
As part of this change, we remove stale odev and that is determined
based on its modTs value.